requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
Pillow
easyocr
transformers
huggingface-hub
sentencepiece
accelerate
scipy
